Home

BanditOpt.jl

BanditOpt.jl is a julia package to make it easy to experiment with Multi-Armed Bandit problems.

BanditOpt — Module.

BanditOpt is a Julia package for experimenting with various Multi-Armed Bandit algorithms.

source

Package Features

This package implements popular MAB algorithms and arm models along with a clean API to easily prototype and evaluate new algorithms wit popular algorithms in literature. The library also aims at reproducible research and provides implementation for a wide set of bandit settings available in Bandit literature.

Manual Outline

Arms

ArmBase

StationaryArmBase

NonstationaryArmBase

Agents

AgentBase

StationaryAgentBase

NonStationaryAgentBase

Stationary Bandit Agents

ϵ-Greedy Agent

ϵₙ-Greedy Agent

Non-stationary Arm Models

Getting Started

Installation

Usage

Accessing documentation from REPL

Stationary Arm Models

Bernoulli Arm

Normal Arm (Gaussian Arm)

BanditOpt.jl

Package Features

Manual Outline