Agents

Agents

This section discusses the common features of agents (algorithms) implemented in this packages.

BanditOpt.Agents — Module.

BanditOpt.Agents module provides implementations of various multi-arm bandit algorithms and an unified API to access those algorithms. This includes base types for implementing various types of agents and the common interface.

source

AgentBase

BanditOpt.Agents.AgentBase — Type.

AgentBase

AgentBase is the base type for all MAB Algorithms

source

Common Interface

get_arm_index()

BanditOpt.Agents.get_arm_index — Function.

get_arm_index( agent::AgentBase )

Gets the index of next arm to pull.

Example

source

update_reward!()

BanditOpt.Agents.update_reward! — Function.

update_reward!( agent::AgentBase, r::Real )

Updates the reward to bandit algorithm agent. r must a real number within valid range.

source

reset!()

BanditOpt.Agents.reset! — Function.

reset!( agent::AgentBase )

Resets the internal statistics of the bandit algorithm, except for the number of arms and other algorithm specific parameters. reset!() will make agent as fresh as it was first created.

source

info_str()

BanditOpt.Agents.info_str — Function.

info_str( agent::AgentBase, latex::Bool = false )

Return a information string about the agent and it's parameters. If flag latex is set to true, then the returned string will be a compactable latex string.

source