Stationary Bandit Agents

Stationary Bandit Agents

This page lists stationary bandit agents available in the package.

ϵ-Greedy Agent

BanditOpt.Agents.epsGreedy — Type.

epsGreedy( noOfArms, ϵ )

Implements constant exploration ϵ-greedy agent. noOfArms is the number of arms to pick from and ϵ is the exploration factor.

source

ϵₙ-Greedy Agent

BanditOpt.Agents.epsNGreedy — Type.

epsNGreedy( noOfArms, c , d )

Implementats decaying exploration factor ϵ-greedy agent. noOfArms is the number of of options, c and d are algorithm dependent parameters.

Reference: Auer, P., Bianchi, N. C., & Fischer, P. (2002). Finite time analysis of the multiarmed bandit problem. Machine Learning, 47, 235–256.

source