Non-stationary Arm Models

Non-stationary Arm Models

This section discusses the non-stationary arm models available in package.

Pulse Arm

BanditOpt.Arms.Pulse — Type.

Pulse( period, changePoint, holdDuration [,isRestless] )

Pulse creates a pulse reward arm which switches between 0 and 1. period(Int) is the period of the pulse wave, changePoint (Int) is the point at which the pulse should switch from 0 to 1 from the start of period. holdDuration (Int) is the time for which the pulse should stay 1. isRestless (bool) sets whether the arm is restless or not.

source

Sinusoidal Arm

BanditOpt.Arms.Sinusoidal — Type.

Sinusoidal( period[, isRestless = true [, offset = 2*π*rand()]] )

Sinusoidal creates an arm which sinusoidally changes between 0 and 1 with period. isRestless (bool) sets whether the arm is restless or not. offset sets the offset for the arm (set to a random offset by default).

source

Square Arm

BanditOpt.Arms.Square — Type.

Square( period::Int, chnagePoints::Dict{Int,Float64}[,isRestless=true] )

Square creates an arm which switches value at specified instances. Starts from a default value of 0 and changes to values as per the dictionary of changePoints.

arm1 = Arms.Square( 50, Dict(10=>0.5,25=>0.20,40=>1.0) )

source

Variational Arm

BanditOpt.Arms.Variational — Type.

Variational( variation, period[, isRestless=true[,offset=2π*rand()]] )

Variational creates a nonstationary arm as mentioned in Besbes, O., Gur, Y., & Zeevi, A. (2014). Optimal Exploration-Exploitation in a Multi-Armed-Bandit Problem with Non-stationary Rewards, 1–20.

source