Stationary Arm Models

Stationary Arm Models

This section discusses the stationary arm models available in package.

Bernoulli Arm

BanditOpt.Arms.Bernoulli — Type.

Bernoulli( p )

Bernoulli creates a Bernoulli arm with success probability p.

source

Beta Arm

BanditOpt.Arms.Beta — Type.

Beta( α, β )

Beta creates a Beta arm with parameters α and β.

source

Exponential Arm

BanditOpt.Arms.Exponential — Type.

Exponetial( θ )

Expoenential creates a Expoenentially distributed arm with mean reward θ

source

Normal Arm (Gaussian Arm)

BanditOpt.Arms.Normal — Type.

Normal( μ, σ )

Creates an arm with reward distributed as Normal random varaible with mean μ and standard deviation σ.

source