Stationary Arm Models

Stationary Arm Models

This section discusses the stationary arm models available in package.

Bernoulli Arm

BanditOpt.Arms.Bernoulli — Type.

Bernoulli( p )

Bernoulli creates a Bernoulli arm with success probability p.

source

Normal Arm (Gaussian Arm)

BanditOpt.Arms.Normal — Type.

Normal( μ, σ )

Creates an arm with reward distributed as Normal random varaible with mean μ and standard deviation σ.

source